requirements.txt
# === ОСНОВНЫЕ ЗАВИСИМОСТИ ===

# Веб-фреймворк
fastapi==0.135.1

# ASGI сервер
uvicorn==0.41.0

# База данных и ORM
sqlalchemy==2.0.48
psycopg2-binary==2.9.11          # PostgreSQL драйвер

# Конфигурация и переменные окружения
python-dotenv==1.2.2

# Валидация данных и сериализация
pydantic==2.12.5
pydantic-settings==2.1.0         # если используете

# === АУТЕНТИФИКАЦИЯ И БЕЗОПАСНОСТЬ ===

# Обработка форм и файлов
python-multipart==0.0.20         # более новая версия

# Хеширование паролей
passlib==1.7.4
bcrypt==5.0.0                    # для passlib

# JWT токены
python-jose==3.5.0               # с криптографией
# или python-jose[cryptography]==3.5.0

# === ТЕСТИРОВАНИЕ ===

# Основной фреймворк тестирования
pytest==9.0.2
pytest-asyncio==1.3.0            # для асинхронных тестов
pytest-cov==7.0.0                # для измерения покрытия кода

# HTTP клиент для тестирования API
httpx==0.28.1

# Асинхронная поддержка
anyio==4.12.1

# Генерация тестовых данных
faker==40.8.0

# === ФРОНТЕНД (Qt) ===

# PySide6 для GUI
PySide6==6.10.2

# HTTP клиент для API запросов из GUI
requests==2.32.5